+++ generator/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## 9.1.6
+
+- Introduced CallAdapters, This feature allows adaptation of a Call with return type R into the type of T. 
+  e.g. Future<User> to Future<Result<User>>
+
+  Code Example:
+
+```dart
+  class MyCallAdapter<T> extends CallAdapterInterface<Future<T>, Future<Either<ApiError, T>>> {
+    @override
+    Future<Either<ApiError, T>> adapt(Future<T> Function() call) async {
+      try {
+        final response = await call();
+        return Either.right(response);
+      }
+      catch (e) {
+        return Either.left(ApiError(e))
+      }
+    }
+  }
+  
+  @RestApi()
+  abstract class RestClient {
+    factory RestClient(Dio dio, {String? baseUrl}) = _RestClient;
+
+    @UseCallAdapter(MyCallAdapter)
+    @GET('/')
+    Future<User> getTasks();
+  }
+```
+
 ## 9.1.5
 
 - Add support for nested object of non-primitive types in `TypedExtras`.